Create ConsoleManager class, refactor SaladBuilder, add test for SaladBuilder, redid SearchingVegetablesTest.

+++ trunk/tasks/VitaliHardzeyenka/ChiefCooker/src/com/epam/training/ChiefCooker/Constants/CucurbitaceaeConstants.java
@@ -1,6 +1,6 @@
 package com.epam.training.ChiefCooker.Constants;
 
 public class CucurbitaceaeConstants {
-	public final static double COEFFICIENT_WEIGHT_OF_PEEL = (double) 0.1;
-	public final static double COEFFICIENT_LIQUID = (double) 0.3;	
+	public final static double COEFFICIENT_WEIGHT_OF_PEEL = 0.1d;
+	public final static double COEFFICIENT_LIQUID = 0.3d;	
 }
